% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brandel.R
\name{brandel}
\alias{brandel}
\title{brandel}
\usage{
brandel(h, diam)
}
\arguments{
\item{h}{is an acctual measured height}

\item{diam}{as an acctual measured diameter}
}
\description{
brandel
}
\examples{
h = 10
diam = 30
brandel(h, diam)
}

