\name{foo}
\alias{foo}
\docType{data}
\title{Example
}
\description{
It is just an example that it can be use
}
\usage{data("foo")}
\format{
  A data frame with 2 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(foo)
str(foo)
}
\keyword{datasets}
