% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/si_contorta.R
\name{si_cont}
\alias{si_cont}
\title{SI_contorta}
\usage{
si_cont(h, a, a0 = 25, a2 = 100)
}
\arguments{
\item{h}{is an acctual measured height}

\item{a}{as an acctual measured age}

\item{a0}{is a special parameter}

\item{a2}{is site index age (SI100 as default)}
}
\description{
SI_contorta
}
\examples{
h=10
a=3
si_cont(h, a)
}

