\name{sverfor}
\alias{sverfor-package}
\docType{package}
\title{
It calculates forest and tree variables for forest and trees in Sweden
}
\description{

~~ In verison 1.0 only brandel function for Norway spruce (Picea abies L. Karst) in southern Sweden and site index for
lodgepole pine in northern Sweden are available ~~
}
\keyword{brandel}
